naive_rag/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pydantic>=2.0.0 